Get Travis CI passing consistently + SwiftWebSocket

+++ ably-ios/ARTStatus.h
@@ -46,11 +46,11 @@ FOUNDATION_EXPORT NSString *const ARTAblyErrorDomain;
 @property (readonly, assign, nonatomic) int code;
 
 // FIXME: use NSInteger instead int (don't know what kind of processor architecture your code might run)
-- (ARTErrorInfo *)setCode:(int) code message:(NSString *) message;
-- (ARTErrorInfo *)setCode:(int) code status:(int) status message:(NSString *) message;
+- (ARTErrorInfo *)setCode:(NSInteger) code message:(NSString *) message;
+- (ARTErrorInfo *)setCode:(NSInteger) code status:(NSInteger) status message:(NSString *) message;
 
-+ (ARTErrorInfo *)createWithCode:(int)code message:(NSString *)message;
-+ (ARTErrorInfo *)createWithCode:(int)code status:(int)status message:(NSString *)message;
++ (ARTErrorInfo *)createWithCode:(NSInteger)code message:(NSString *)message;
++ (ARTErrorInfo *)createWithCode:(NSInteger)code status:(NSInteger)status message:(NSString *)message;
 // FIXME: base NSError
 + (ARTErrorInfo *)createWithNSError:(NSError *)error;
